Add files via upload
The App Rating prediction has been a topic of interest in recent years. The challenging part was to identify a computationally low and cost-efficient algorithm that provided decent results. Target class imbalance is a major issue in real-world datasets, and it has been handled using a SMOTE technique in combination with an optimised Random Forest model using the GridSearchCV method to identify ideal hyper-parameters. It provided the best results and a weighted F1 score of ‘0.86’. Using the Variable of Importance plot, a total of 10 decision-making features for app rating prediction were identified and the business values associated with it are summarised.